Cast ARGB literals to Int in the luma unit test.
0xFF804020 does not fit in a Kotlin Int, so IntArray init failed to
compile; toInt() keeps the packed channel bits.

+++ core/ai/src/test/java/com/vitacut/core/ai/AnalyzersTest.kt
@@ -284,7 +284,7 @@ class TemplateMatcherTest {
 
     @Test
     fun `toGray converts luma`() {
-        val pixels = IntArray(4) { 0xFF804020 } // r=128 g=64 b=32
+        val pixels = IntArray(4) { 0xFF804020.toInt() } // r=128 g=64 b=32
         val gray = TemplateMatcher.toGray(pixels, 2, 2)
         // 128*77 + 64*150 + 32*29 = 9856+9600+928 = 20384 → shr 8 = 79
         assertEquals(79, gray[0].toInt() and 0xFF)
